+++ FaragonEngine/src/Platform/Windows/WindowsInput.h
@@ -12,9 +12,9 @@ namespace FaragonEngine
 
 		virtual bool IsKeyPressedImpl(int key) override;
 		virtual bool IsMouseButtonPressedImpl(int button) override;
-		virtual std::pair<float, float> GetMousePositionImpl() override;
-		virtual float GetMouseXImpl() override;
-		virtual float GetMouseYImpl() override;
+		virtual std::pair<double, double> GetMousePositionImpl() override;
+		virtual double GetMouseXImpl() override;
+		virtual double GetMouseYImpl() override;
 	};
 }
